+++ tests/conftest.py
@@ -1,0 +1,9 @@
+from pathlib import Path
+
+import pytest
+
+
+@pytest.fixture
+def test_file_path():
+    return Path(__file__).parent / "files" / "rwservlet.pdf"
+
