+++ tests/test_tika.py
@@ -14,56 +14,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from http import HTTPStatus
-from pathlib import Path
-
-import tika.parser
 import tika.tika
-
-TEST_FILE_PATH = Path(__file__).parent / "files" / "rwservlet.pdf"
-
-
-def test_remote_pdf():
-    """parse remote PDF"""
-    assert tika.parser.from_file(
-        "https://upload.wikimedia.org/wikipedia/commons/4/42/Article_feedback_flow_B_-_Thank_editors.pdf")
+from tika import parser
 
 
-def test_remote_html():
-    """parse remote HTML"""
-    assert tika.parser.from_file("http://nossl.sh")
-
-
-def test_remote_mp3():
-    """parse remote mp3"""
-    assert tika.parser.from_file(
-        "https://archive.org/download/Ainst-Spaceshipdemo.mp3/Ainst-Spaceshipdemo.mp3")
-
-
-def test_remote_jpg():
-    """parse remote jpg"""
-    assert tika.parser.from_file(
-        "https://upload.wikimedia.org/wikipedia/commons/b/b7/X_logo.jpg")
-
-
-def test_local_binary():
-    """parse file binary"""
-    with open(TEST_FILE_PATH, "rb") as file_obj:
-        assert tika.parser.from_file(file_obj)
-
-
-def test_local_buffer():
-    response = tika.parser.from_buffer("Good evening, Dave")
-    assert response["status"] == HTTPStatus.OK
-
-
-def test_local_path():
-    """parse file path"""
-    assert tika.parser.from_file(str(TEST_FILE_PATH))
-
-
-def test_kill_server():
+def test_kill_server(test_file_path):
     """parse some file then kills server"""
-    with open(TEST_FILE_PATH, "rb") as file_obj:
+    with open(test_file_path, "rb") as file_obj:
         tika.parser.from_file(file_obj)
     assert tika.tika.killServer() is None
